requirements1.txt
appdirs==1.4.3
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.1
django-adminlte2==0.1.6
django-bootstrap3==9.1.0
django-crispy-forms==1.6.1
django-registration-redux==1.6
djangorestframework==3.7.3
idna==2.6
oauthlib==2.0.6
olefile==0.44
packaging==16.8
Pillow==4.1.1
psycopg2==2.7.1
PyJWT==1.5.3
pyparsing==2.2.0
python-openid==2.2.5
python-social-auth==0.2.1
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==2.0.0
social-auth-core==1.5.0
urllib3==1.22
